refactor move mock data from fixtures to mock folder add tag to table cell text add more mock data fields
