+++ ui/src/App.js
@@ -1,5 +1,5 @@
 import Dashboard from "./Components/Dashboard/Dashboard";
-import { FEATURE_FLAGS } from "./test/Fixtures/featureFlags";
+import { FEATURE_FLAGS, FEATURE_FLAGS_HEADERS} from "./Mock/featureFlags";
 
 function App() {
   return (
@@ -8,7 +8,7 @@ function App() {
     >
       <Dashboard 
         featureFlags={FEATURE_FLAGS} 
-        headers={['featureFlag', 'repository']}
+        headers={FEATURE_FLAGS_HEADERS}
       />
     </main>
   );
