+++ ui/src/Mock/featureFlags.js
@@ -1,0 +1,34 @@
+export const FEATURE_FLAGS = [
+  {
+    id: 1,
+    featureFlag: 'Dark Mode',
+    repository: 'Member Site',
+    enabled: true,
+    createdBy: 'Rohan Raj',
+    createdAt: 1657868257983,
+    modifiedAt: 1657868757983,
+  },
+  {
+    id: 2,
+    featureFlag: 'Cloudinary',
+    repository: 'Welcome Site',
+    enabled: false,
+    createdBy: 'Pallab Sonowal',
+    createdAt: 1657868057983,
+    modifiedAt: 1657863757983,
+  },
+  {
+    id: 3,
+    featureFlag: 'Get Intro',
+    repository: 'Member Site',
+    enabled: true,
+    createdBy: 'Rahul',
+    createdAt: 1657868240983,
+    modifiedAt: 1657868807983,
+  },
+]
+
+export const FEATURE_FLAGS_HEADERS = [
+  'featureFlag', 
+  'repository',
+]
