--- ui/src/test/Fixtures/featureFlags.js
@@ -1,20 +1,0 @@
-export const FEATURE_FLAGS = [
-  {
-    id: 1,
-    featureFlag: 'Dark Mode',
-    repository: 'Member Site',
-    enabled: true,
-  },
-  {
-    id: 2,
-    featureFlag: 'Cloudinary',
-    repository: 'Welcome Site',
-    enabled: false,
-  },
-  {
-    id: 3,
-    featureFlag: 'Get Intro',
-    repository: 'Member Site',
-    enabled: true,
-  },
-]
